fix(lottery): use clearTimeout
